Rewrote language defs using Pattern Library, other refactorings
